ms/mm: cma: fix incorrect type conversion for size during dma allocation
This was found during userspace fuzzing test when a large size dma cma
allocation is made by driver(like ion) through userspace.

  show_stack+0x10/0x1c
  dump_stack+0x74/0xc8
  kasan_report_error+0x2b0/0x408
  kasan_report+0x34/0x40
  __asan_storeN+0x15c/0x168
  memset+0x20/0x44
  __dma_alloc_coherent+0x114/0x18c

Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Marek Szyprowski <m.szyprowski@samsung.com>

=========================
(cherry picked from commit 67a2e213e7e937c41c52ab5bc46bf3f4de469f6e)

Conflicts:
	include/linux/cma.h
	mm/cma.c

Backport changes: dropped hunks related to cma_alloc()
as it does not exist.

https://jira.sw.ru/browse/PSBM-73053

CVE-2017-9725 kernel: Incorrect type conversion for size during dma allocation

A flaw was found in the Linux kernels implementation of DMA memory allocation.

When allocating a DMA buffer, a section of memory is allocated and then set to
zeros.  The size parameter of allocation was truncated due to an incorrect
casting when the allocation function is called.  During clearing the allocation
used an untruncated value as the size to clear and would "zero" a larger
section of kernel memory than was allocated, possibly corrupting memory and
allowing for privilege escalation.

At this time Red Hat Product Security believes that there is no direct control
of the size parameter used in this function in Red Hat kernels.

+++ drivers/base/dma-contiguous.c
@@ -272,7 +272,7 @@ err:
  * global one. Requires architecture specific get_dev_cma_area() helper
  * function.
  */
-struct page *dma_alloc_from_contiguous(struct device *dev, int count,
+struct page *dma_alloc_from_contiguous(struct device *dev, size_t count,
 				       unsigned int align)
 {
 	unsigned long mask, pfn, pageno, start = 0;
